requirements.txt
# Python dependencies for the supported full-production dashboard install.
# `requirements.txt` is the authoritative uv install contract for the
# Ubuntu 24.04 + CUDA 13 + Python 3.10 runtime path.
# The bundled Diff_GIPC/libuipc wheel is installed by the setup scripts.
# Repo-local extern trees and native binaries are provisioned by the setup
# harness, but the main production Python dependency set lives here.
--extra-index-url https://download.pytorch.org/whl/nightly/cu130
--extra-index-url https://pypi.ngc.nvidia.com
accelerate==1.13.0
depth-pro @ git+https://github.com/apple/ml-depth-pro.git
diffusers==0.37.0
einops==0.8.2
huggingface-hub==1.6.0
libigl==2.6.2
matplotlib==3.10.8
ninja==1.13.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.23.2
open3d==0.19.0
openai==2.23.0
opencv-python==4.11.0.86
Pillow==12.1.1
plyfile==1.1.3
polyscope==2.6.0
pybind11==3.0.2
pygltflib==1.16.5
pymeshfix==0.18.0
pymeshlab==2025.7.post1
rembg==2.0.69
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
tiktoken==0.12.0
timm==1.0.25
trimesh==4.11.2
torch==2.13.0.dev20260421+cu130
torchvision==0.27.0.dev20260421+cu130
transformers==5.3.0
xatlas==0.0.11
